+++ CHANGELOG.md
@@ -2,14 +2,26 @@
 
 [Unreleased changes](./UNRELEASED.yaml).
 
+# 2020-04-30
+
+**Added:**
+
+  - bpk-component-loading-button: 3.0.49 => 3.1.0
+    - Added a `iconPosition` prop which allows consumers to control whether the icon is shown on the leading or trailing side of the button.
+
+**Fixed:**
+
+  - bpk-component-flare: 1.0.47 => 1.0.48
+    - Fixed an issue that caused the flare to disappear in RTL.
+
 # 2020-04-20 - RTL fixes
 
 **Fixed:**
 
-  - bpk-component-rating:
+  - bpk-component-rating: 2.0.45 => 2.0.46
     - Fixed a couple of minor layout issues in RTL.
 
-  - bpk-component-star-rating:
+  - bpk-component-star-rating: 2.1.43 => 2.1.44
     - Fixed a bug that caused a strange animation when selecting a star rating in RTL mode.
 
 # 2020-03-09 - New `primary` chip type.